requirements.txt
requests>=2.31.0
Pillow>=10.0.0
duckduckgo-search>=4.1.1
beautifulsoup4>=4.12.0
urllib3>=2.0.0
tldextract>=5.1.1 